+++ src/Tests/Dfe.Complete.Api.Tests.Integration/Controllers/TasksDataController/UpdateCompleteNotificationOfChangeTaskTests.cs
@@ -1,0 +1,99 @@
+using AutoFixture;
+using AutoFixture.Xunit2;
+using Dfe.Complete.Api.Tests.Integration.Customizations;
+using Dfe.Complete.Client.Contracts;
+using Dfe.Complete.Domain.Entities;
+using Dfe.Complete.Infrastructure.Database;
+using Dfe.Complete.Tests.Common.Constants;
+using GovUK.Dfe.CoreLibs.Testing.AutoFixture.Attributes;
+using GovUK.Dfe.CoreLibs.Testing.Mocks.WebApplicationFactory;
+using Microsoft.EntityFrameworkCore;
+using System.Security.Claims;
+
+namespace Dfe.Complete.Api.Tests.Integration.Controllers.TasksDataController
+{
+    public class UpdateCompleteNotificationOfChangeTaskTests
+    {
+        [Theory]
+        [CustomAutoData(
+            typeof(CustomWebApplicationDbContextFactoryCustomization),
+            typeof(ConversionTaskDataCustomization))]
+        public async Task UpdateCompleteNotificationOfChangeTaskAsync_ShouldUpdate_ConversionTaskData(
+            CustomWebApplicationDbContextFactory<Program> factory,
+            ITasksDataClient tasksDataClient,
+            [Frozen]
+            IFixture fixture)
+        {
+            // Arrange
+            factory.TestClaims = [new Claim(ClaimTypes.Role, ApiRoles.ReadRole), new Claim(ClaimTypes.Role, ApiRoles.UpdateRole), new Claim(ClaimTypes.Role, ApiRoles.WriteRole)];
+
+            var dbContext = factory.GetDbContext<CompleteContext>();
+
+            var taskData = fixture.Create<ConversionTasksData>();
+            dbContext.ConversionTasksData.Add(taskData);
+
+            await dbContext.SaveChangesAsync();
+
+            var command = new UpdateCompleteNotificationOfChangeTaskCommand
+            {
+                TaskDataId = new TaskDataId { Value = taskData.Id.Value },
+                TellLocalAuthority = true,
+                CheckDocument = true,
+                SendDocument = true,
+            };
+            
+            // Act
+            await tasksDataClient.UpdateCompleteNotificationOfChangeTaskAsync(command, default);
+
+            // Assert
+            dbContext.ChangeTracker.Clear();
+            var existingTaskData = await dbContext.ConversionTasksData.SingleOrDefaultAsync(x => x.Id == taskData.Id);
+            Assert.NotNull(existingTaskData);
+            Assert.True(existingTaskData.CompleteNotificationOfChangeTellLocalAuthority);
+            Assert.True(existingTaskData.CompleteNotificationOfChangeCheckDocument);
+            Assert.True(existingTaskData.CompleteNotificationOfChangeSendDocument);
+            Assert.Null(existingTaskData.CompleteNotificationOfChangeNotApplicable);
+        }
+
+        [Theory]
+        [CustomAutoData(
+            typeof(CustomWebApplicationDbContextFactoryCustomization),
+            typeof(ConversionTaskDataCustomization))]
+        public async Task UpdateCompleteNotificationOfChangeTaskAsync_ShouldUpdateNotApplicableOnly(
+            CustomWebApplicationDbContextFactory<Program> factory,
+            ITasksDataClient tasksDataClient,
+            IFixture fixture)
+        {
+            // Arrange
+            factory.TestClaims = [new Claim(ClaimTypes.Role, ApiRoles.ReadRole), new Claim(ClaimTypes.Role, ApiRoles.UpdateRole), new Claim(ClaimTypes.Role, ApiRoles.WriteRole)];
+
+            var dbContext = factory.GetDbContext<CompleteContext>();
+
+            var taskData = fixture.Create<ConversionTasksData>();
+            dbContext.ConversionTasksData.Add(taskData);
+
+            await dbContext.SaveChangesAsync();
+
+            var command = new UpdateCompleteNotificationOfChangeTaskCommand
+            {
+                TaskDataId = new TaskDataId { Value = taskData.Id.Value },
+                NotApplicable = true,
+                TellLocalAuthority = true,
+                CheckDocument = true,
+                SendDocument = true,
+            };
+
+            // Act
+            await tasksDataClient.UpdateCompleteNotificationOfChangeTaskAsync(command, default);
+
+            // Assert
+            dbContext.ChangeTracker.Clear();
+            var existingTaskData = await dbContext.ConversionTasksData.SingleOrDefaultAsync(x => x.Id == taskData.Id);
+            Assert.NotNull(existingTaskData);
+            Assert.Null(existingTaskData.CompleteNotificationOfChangeTellLocalAuthority);
+            Assert.Null(existingTaskData.CompleteNotificationOfChangeCheckDocument);
+            Assert.Null(existingTaskData.CompleteNotificationOfChangeSendDocument);
+            Assert.True(existingTaskData.CompleteNotificationOfChangeNotApplicable);
+        }
+    }
+}